+++ my-app/images/desktop.ini
@@ -1,0 +1,4 @@
+[ViewState]
+Mode=
+Vid=
+FolderType=Generic
