+++ my-app/my-app/package.json
@@ -10,14 +10,19 @@
     "@mapbox/polyline": "^1.1.0",
     "@nomoreanalog/react-scripts-eslintless": "2.1.10",
     "babel-plugin-styled-components": "^1.10.7",
+    "bootstrap": "^4.5.0",
+    "geojson": "^0.5.0",
     "link": "^0.1.5",
     "mapbox-gl": "^1.10.0",
     "native-base": "^2.13.12",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-native": "^0.62.2",
+    "react-native-maps": "^0.27.1",
     "react-native-svg": "^12.1.0",
-    "react-native-vector-icons": "^6.6.0"
+    "react-native-vector-icons": "^6.6.0",
+    "react-native-web": "^0.12.2",
+    "react-router-dom": "^5.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -30,5 +35,9 @@
     "not dead",
     "not ie <= 11",
     "not op_mini all"
-  ]
+  ],
+  "devDependencies": {
+    "@babel/cli": "^7.10.1",
+    "@babel/core": "^7.10.2"
+  }
 }
