--- my-app/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "my-app",
-  "version": "1.0.0",
-  "description": "Towtug",
-  "main": "server.js",
-  "dependencies": {
-    "react-router-dom": "^5.1.2"
-  },
-  "devDependencies": {},
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "server.js"
-  },
-  "author": "",
-  "license": "ISC"
-}
